+++ site/themes/default/views/templates/about.php
@@ -2,5 +2,5 @@
 <?php Themes::view('partials/head')->display() ?>
 <h1><?= $entry['title'] ?></h1>
 <?= $entry['content'] ?>
-<img src="<?= $entry['base_url'] ?>/<?= $entry['slug'] ?>/<?= $entry['image'] ?>">
+<img src="<?= Images::getImageUrl($entry['slug'] .'/'. $entry['image'], ['w' => '670', 'dpr' => '2']) ?>">
 <?php Themes::view('partials/footer')->display() ?>